+++ persistence/memorypersistence/process.go
@@ -46,47 +46,20 @@ func (v *validator) VisitSaveProcessInstance(
 	}
 }
 
-// VisitRemoveProcessInstance returns an error if a "RemoveProcessInstance"
-// operation can not be applied to the database.
-func (v *validator) VisitRemoveProcessInstance(
-	ctx context.Context,
-	op persistence.RemoveProcessInstance,
-) error {
-	inst := op.Instance
-	key := instanceKey{inst.HandlerKey, inst.InstanceID}
-
-	if x, ok := v.db.process.instances[key]; ok {
-		if inst.Revision == x.Revision {
-			return nil
-		}
-	}
-
-	return persistence.ConflictError{
-		Cause: op,
-	}
-}
-
 // VisitSaveProcessInstance applies the changes in a "SaveProcessInstance"
 // operation to the database.
 func (c *committer) VisitSaveProcessInstance(
 	ctx context.Context,
 	op persistence.SaveProcessInstance,
 ) error {
-	c.db.process.save(op.Instance)
-	return nil
-}
-
-// VisitRemoveProcessInstance applies the changes in a "RemoveProcessInstance"
-// operation to the database.
-func (c *committer) VisitRemoveProcessInstance(
-	ctx context.Context,
-	op persistence.RemoveProcessInstance,
-) error {
 	inst := op.Instance
-	key := instanceKey{inst.HandlerKey, inst.InstanceID}
+	c.db.process.save(inst)
 
-	c.db.process.remove(key)
-	c.db.queue.removeTimeoutsByProcessInstance(key)
+	if inst.HasEnded {
+		key := instanceKey{inst.HandlerKey, inst.InstanceID}
+		c.db.queue.removeTimeoutsByProcessInstance(key)
+	}
+
 	return nil
 }
 
@@ -105,9 +78,4 @@ func (db *processDatabase) save(inst persistence.ProcessInstance) {
 
 	inst.Revision++
 	db.instances[key] = inst
-}
-
-// remove removes the process instance with the given key.
-func (db *processDatabase) remove(key instanceKey) {
-	delete(db.instances, key)
 }